improve downloaded plugin caching - check if a newer compatible version is available on the marketplace
We usually don’t specify the plugin version explicitly. Instead, the latest compatible version is downloaded automatically based on the IntelliJ version. However, if the plugin is released with some dealy after the IDE, it can cause issues with build reproducibility: one machine might cache an older plugin version if the project is imported just after IDE release, while another machine might use the newer version if by that time a newer plugin is released. This can lead to differences in builder behavior on two different machines. To prevent this, we improved our plugin caching logic: instead of just checking that the plugin directory exists, we now compare the cached plugin file name with the latest one available remotely. This is only done for those plugins that are known to be downloaded (non-bundled).

Note, to understand that the plugin is downloaded, we store information in the plugin index file.
It's currently stored in the "downloadedPluginFileName" field.
The name itself is effectively not used at the moment, and it would be enough to store just a boolean value.
But it serves as an additional hint that can be useful during the debugging.

ATTENTION: if you delete the plugin index file, then the information about the plugin being downloaded will be lost and no newer versions will be checked